Split mDNS headers to C and C++
For some reason, this was missed?
