+++ wpinet/src/main/native/include/wpi/net/MulticastServiceAnnouncer.h
@@ -7,80 +7,7 @@
 #include <stdint.h>
 
 #ifdef __cplusplus
-
-#include <memory>
-#include <span>
-#include <string>
-#include <string_view>
-#include <utility>
-
-namespace wpi::net {
-
-class MulticastServiceAnnouncer {
- public:
-  /**
-   * Creates a MulticastServiceAnnouncer.
-   *
-   * @param serviceName service name
-   * @param serviceType service type
-   * @param port port
-   * @param txt txt
-   */
-  MulticastServiceAnnouncer(
-      std::string_view serviceName, std::string_view serviceType, int port,
-      std::span<const std::pair<std::string, std::string>> txt);
-
-  /**
-   * Creates a MulticastServiceAnnouncer.
-   *
-   * @param serviceName service name
-   * @param serviceType service type
-   * @param port port
-   * @param txt txt
-   */
-  MulticastServiceAnnouncer(
-      std::string_view serviceName, std::string_view serviceType, int port,
-      std::span<const std::pair<std::string_view, std::string_view>> txt);
-
-  /**
-   * Creates a MulticastServiceAnnouncer.
-   *
-   * @param serviceName service name
-   * @param serviceType service type
-   * @param port port
-   */
-  MulticastServiceAnnouncer(std::string_view serviceName,
-                            std::string_view serviceType, int port);
-
-  ~MulticastServiceAnnouncer() noexcept;
-
-  /**
-   * Starts multicast service announcer.
-   */
-  void Start();
-
-  /**
-   * Stops multicast service announcer.
-   */
-  void Stop();
-
-  /**
-   * Returns true if there's a multicast service announcer implementation.
-   *
-   * @return True if there's a multicast service announcer implementation.
-   */
-  bool HasImplementation() const;
-
-  struct Impl;
-
- private:
-  std::unique_ptr<Impl> pImpl;
-};
-
-}  // namespace wpi::net
-
 extern "C" {
-
 #endif
 
 typedef unsigned int WPI_MulticastServiceAnnouncerHandle;  // NOLINT
